+++ fluent-fallback/src/types.rs
@@ -16,13 +16,13 @@ impl<'l> From<&'l str> for L10nKey<'l> {
     }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct L10nAttribute<'l> {
     pub name: Cow<'l, str>,
     pub value: Cow<'l, str>,
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct L10nMessage<'l> {
     pub value: Option<Cow<'l, str>>,
     pub attributes: Vec<L10nAttribute<'l>>,
